Added local.mk.sample - an example of modifying the make variables
The sample can be copied by the user to local.mk, and then edited to
modify F90C, F90GLAGS, etc. An example is provide for the Intel
Fortran compiler on Linux x64. The cfitsio link options are
particularly finickety on Fedora. I could never get simply "-lcfitsio"
to work.